move deprecated_alias to lib_losoto.py
