+++ losoto/h5parm.py
@@ -9,32 +9,12 @@ import numpy as np
 import tables
 import losoto._version
 from losoto._logging import logger as logging
-
+from losoto.lib_losoto import deprecated_alias
 # check for tables version
 if int(tables.__version__.split('.')[0]) < 3:
     logging.critical('pyTables version must be >= 3.0.0, found: '+tables.__version__)
     sys.exit(1)
 
-# fancy backwards compatibility of keywords: allow aliases
-# https://stackoverflow.com/questions/49802412/how-to-implement-deprecation-in-python-with-argument-alias#
-def deprecated_alias(**aliases):
-    def deco(f):
-        @functools.wraps(f)
-        def wrapper(*args, **kwargs):
-            rename_kwargs(f.__name__, kwargs, aliases)
-            return f(*args, **kwargs)
-        return wrapper
-    return deco
-
-def rename_kwargs(func_name, kwargs, aliases):
-    for alias, new in aliases.items():
-        if alias in kwargs:
-            if new in kwargs:
-                raise TypeError('{} received both {} and {}'.format(
-                    func_name, alias, new))
-            warnings.warn('{} is deprecated; use {}'.format(alias, new),
-                          DeprecationWarning)
-            kwargs[new] = kwargs.pop(alias)
 
 def openSoltab(h5parmFile, solsetName=None, soltabName=None, address=None, readonly=True):
     """
